+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This release extends support to OneView Rest API version 1800 (OneView v5.30) fo
 - FCOE Network
 - Hypervisor Cluster Profile
 - Hypervisor Manager
+- Interconnects
+- Logical Enclosure
 - Network Set
 - Scope
 - Server Certificate